// src/Models/admin-emp.model.ts
export class Employee {


    employeeId: number = 0;
    employeeFirstName: string = '';
    status: boolean = true;
    employeeLastName: string = '';
    salary: number = 0;
    email: string = '';
    phoneNumber: number = 0;
    userId:number=0;
    userName: string = '';
    password: string = '';

}